RstLexer: improve handling of field lists
The current field list rules only find a field body if it starts on the
same name as the field marker, as shown by this code sample [1] in the
Sphinx docs. This replacement rule covers any field marker (including
colons if escaped) and an optional field body consisting of one or more
paragraphs, starting on either the same line or a following line, as
long as it's indented more than the marker.

[1]: http://www.sphinx-doc.org/en/stable/domains.html#the-javascript-domain

Taken from https://bitbucket.org/birkenfeld/pygments-main/pull-requests/622/rstlexer-improve-handling-of-field-lists/diff